enhancement(editor): Splitting component, extracting EditorPanel logic to its own module to make it more maintainable
